requirements.txt
openai>=1.60.0
python-dotenv>=1.0.1
gigachat>=0.2.0
flask>=3.0.0
